Fix Vite HMR breakage in Docker container
If server.hmr.host is unset, it defaults to localhost.  When Vite server
boots up, the link to javascript assets compiled by Vite prepends with
localhost.  This breaks HMR when Vite dev server is running concurrently
with PHP development server in a Docker container hosted on a remote
Linux server.

This is a bug and it is 100% reproducible.  If server.hmr.host defaults
to localhost due to security reason, it is designed poorly since Vite
server is for development that doesn't need such security.

To fix this bug, set server.hmr.host to a valid link but localhost.  The
link will point to the live Vite server for the assets for now.

Ultimately the link will be represented in an environment variable that
has the value of the host URL.