Add v-model directive to manage greeting property
Use v-model directive, which is a special HTML attribute-like syntax
that binds the value of an HTML element to a data property on Vue
instance.  It's used to create two-way data binding, meaning changes to
the input element will automatically update the corresponding data
property, and vice versa.

The v-model="greeting" directive binds the value of the input element to
the greeting data property.  If the user types something into the input,
the greeting property will be updated automatically.  Similarly, if we
update the greeting property in JavaScript code, the input value will be
updated.

Add the display of how many characters are in the greeting property that
the user enters at a given time.