Add mounted() function to show the effect
The mounted() function is a life cycle hook that is called after the
component has been mounted to the DOM.  This means that the component's
template has been rendered, and the component's data and methods are
ready for use.

With mounted() method, we can use it to initialize any necessary tasks
for the component.  In our example, we change the greeting 4 seconds
after mounted() is called and succeeds.