Add label to nest input element
By nesting the <input> directly inside the <label>, the for and id
attributes are spared for the implicit association.  This nesting is
required as there are multiple label and checkbox combos, making each
checkbox implicitly associated with its label.  Now it's the same by
clicking on either the text or on the checkbox.