Use Javascript object for class
The use of v-bind to class lets expression to assign a list of Tailwind
utilities to it.  The expression is made of Javascript object with the
key of the list and the value of true.  The true value applies the list
to the class.  If the value is false, it doesn't apply the list.