Add processing in props
Within data() method, processing was an object of data.  As props being
built, use processing as an object of props.  Declare it as boolean with
the default value of false.  Both processing objects can't exist at the
same time and get rid of data() method entirely.

The existing v-bind of disabled will continue to work as it is.  In the
context of v-bind, either processing remains to be the same to v-bind.

With the default value of processing set to false, disabled doesn't show
any more.  It was true and disabled did show.