Add jobList component for to-do or completed jobs
Loop through to-do or completed job list is almost identical.  Use one
component to replace both.  In each list, use props for dynamic values
to build a generic component for either to-do or completed jobs.

Use allJobs in lieu of jobs as the key for the full list of jobs,
regardless of the status.