Use AllOver for the name of imported component
When AllJobs is imported as a component, an identifier must be used to
reference it.  Use the name of AllOver for that.

AllOver must be registered before it can be used.  In the context of
components: { AllOver }, AllOver is a shorthand for 'AllOver: AllOver'.

Now <AllOver></AllOver> is ready for use in template.

Clean up div wrapping around <AllOver></AllOver>.  Use <Head></Head>.