Add space between JobList
Use section node as parent of JobList.  Now space-y-8 will be applied
between each section as a whole rather than between nodes of JobList.
This is a much better approach than the use of div with mt-8.  For
instance, if more JobList are added, the space will be automatically
appended.  But with the use of div, more div node is required.

Use v-show to section to hide whole JobList should it not show.

Fix formatting.