Add a form for adding a new job
Add a form below the lists of jobs with Add button to submit a new job.
Build the basic components of input and button for the form.  Make the
form look good in alignment with the job lists.  Lay down the basic work
of the form for the next phase of development using vue.