Add a new job to all jobs
Construct a new job object and push the job to the array of allJobs.

The user enters the name of the new job.  If the user doesn't enter the
name, it uses the default value.  The status is set to false and the id
is the size of the original array plus 1.

Set the default new job name to Apply Inertia and Vue.

With the addition of new job implemented, remove alert call.