Use v-model to manage currentTag
Use v-model to set the currentTag value.  In this example it shows
currentTag is managed by modelValue prop internally.  The name of
modelValue is arbitrary and it is this way by default.

As v-model works bidirectionally, this example only shows the part by
v-model of setting currentTag value.

Next, we will see how currentTag value updates tag style.